Adding in more detail.
I think H2 will be a simple way to deal with data persistence for now.  Not really a desirable way, but fine for prototyping like this.

Build has problems, I may have to rethink using this library for UI things.  I was hoping for a friendly way to create simple React, but npm and yarn aren't playing well.